pyproject.toml
[project]
name = "mesh-n-bone"
version = "0.1.0"
description = "Unified tool for mesh generation, multiresolution mesh creation, skeletonization, and analysis."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "David Ackerman", email = "ackermand@janelia.hhmi.org" },
]
license-files = ["LICENSE"]

dependencies = [
  # Core
  "numpy",
  "trimesh>=4.6.8",
  "dask[distributed]==2025.5.1",
  "dask-jobqueue==0.9.0",
  "bokeh>=3.1.0",
  "pyyaml",
  "pandas",

  # Mesh processing
  "dracopy>=1.5.0",
  "pyfqmr>=0.3.0",
  "pymeshlab",

  # Mesh generation (zmesh, cloud-volume)
  "cloud-volume",
  "zmesh",

  # Volumetric data
  "funlib.geometry",
  "tensorstore",
  "zarr>=3.0.8",
  "numba",

  # Skeletonization
  "neuroglancer",
  "fastremap",
  "pybind11-rdp",
  "networkx",

  # Analysis
  "scipy",
  "pyarrow",

  # Geometry
  "shapely",
]

[project.optional-dependencies]
docs = [
  "mkdocs-material",
  "mkdocstrings[python]",
]

[project.urls]
Repository = "https://github.com/janelia-cellmap/mesh-n-bone"
Documentation = "https://janelia-cellmap.github.io/mesh-n-bone/"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
mesh-n-bone = "mesh_n_bone.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["mesh_n_bone*"]
